rename de componentes se reifica la vista de place en el form de registro
